+++ fabrix-mg-derive/src/indexes.rs
@@ -1,6 +1,6 @@
-//！ Indexes
-//！
-//！ Single
+//! Indexes
+//!
+//! Single
 
 use std::str::FromStr;
